注册字眼修改、支付宝URLScheme修改

+++ ZhiMaDi/UserCenter/Order/OrderCommentViewController.swift
@@ -234,12 +234,14 @@ class OrderCommentViewController: UIViewController,UITableViewDelegate,UITableVi
     //MARK: - MWPhotoBrowserDelegate
     func numberOfPhotosInPhotoBrowser(photoBrowser: MWPhotoBrowser!) -> UInt {
         let photos = self.photos[self.photoIndex]
-        return UInt(photos.count)
+        return photos.count
+//        return UInt(photos.count)
     }
     func photoBrowser(photoBrowser: MWPhotoBrowser!, photoAtIndex index: UInt) -> MWPhotoProtocol! {
         let photos = self.photos[self.photoIndex]
         if Int(index) < photos.count {
-            let photo:MWPhoto = MWPhoto(image: photos[Int(index)] as! UIImage)
+            let photo : MWPhoto = MWPhoto(image: photos[index] as! UIImage)
+//            let photo:MWPhoto = MWPhoto(image: photos[Int(index)] as! UIImage)
             return photo
         }
         return nil
